make products work
