+++ core/urls.py
@@ -21,6 +21,7 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('shop.urls')),
+    path('products/', include('products.urls')),
     path('accounts/', include('django.contrib.auth.urls')),
 
 ]
