+++ products/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
+from .models import Product, Category, Feedback
 
 # Register your models here.
+admin.site.register(Product)
+admin.site.register(Category)
+admin.site.register(Feedback)
