+++ products/templates/products/products.html
@@ -1,0 +1,15 @@
+{% extends "shop/base.html" %}
+{% load static %}
+
+{% block css %}
+	
+{% endblock  %}
+
+{% block content %}	
+					<div class="auto-container">
+						<h2> Product list </h2>
+
+								product1
+                                product2
+					</div>
+{% endblock%}
