+++ shop/admin.py
@@ -1,7 +1,4 @@
 from django.contrib import admin
-from .models import Product, Category, Feedback
+
 
 # Register your models here.
-admin.site.register(Product)
-admin.site.register(Category)
-admin.site.register(Feedback)
